feat/SeperateLicenseView
